Heavily improved the AI.

+++ src/main/java/org/nschmidt/abalone/ai/AI.java
@@ -16,7 +16,7 @@ public enum AI {
     
     public static Field bestMove(Field state, Player currentPlayer) {
         Field[] moves = MoveDetector.allMoves(state, currentPlayer);
-        Arrays.sort(moves, (m1, m2) -> Integer.compare(score(m2, currentPlayer), score(m1, currentPlayer)));
+        Arrays.sort(moves, (m1, m2) -> Double.compare(score(m2, currentPlayer), score(m1, currentPlayer)));
         
         if (moves.length == 0) return state;
         if (moves.length == 1) return moves[0];
